+++ scripts/add_libs_to_wheel.sh
@@ -31,7 +31,6 @@ if [[ "$DOCKER_ARCH" == "amd64" ]]; then
 elif [[ "$DOCKER_ARCH" == "arm64" ]]; then
   WHEEL_ARCH=aarch64
 fi
-WHEEL_PLATFORM=manylinux_2_35_${WHEEL_ARCH}
 
 DOCKER_CONTAINER_ID=$(docker create --rm --platform "${DOCKER_PLATFORM}" -w "${PWD}" "${PYTRITON_IMAGE_NAME}" bash -c "sleep 1h")
 docker start "${DOCKER_CONTAINER_ID}"
@@ -42,9 +41,14 @@ docker exec "${DOCKER_CONTAINER_ID}" mkdir -p "$(dirname "${TRITON_LOCAL_DIR}")"
 docker cp "${WHEEL_PATH}" "${DOCKER_CONTAINER_ID}:${WHEEL_PATH}"
 docker cp "${TRITON_LOCAL_DIR}" "${DOCKER_CONTAINER_ID}:${TRITON_LOCAL_DIR}"
 
-# The setuptools are necessary for Python 3.12
-# pyproject.toml also contains the setuptools version so please update it accordingly
-docker exec "${DOCKER_CONTAINER_ID}" pip install auditwheel==5.3.0 patchelf==0.17.2 setuptools==69.2.0
+# The setuptools and packaging are necessary for Python 3.12
+# pyproject.toml also contains the setuptools and packaging versions so please update it accordingly
+docker exec "${DOCKER_CONTAINER_ID}" pip install auditwheel==5.3.0 patchelf==0.17.2 setuptools==69.2.0 packaging==24.0
+
+docker cp "${SCRIPTS_DIR}/manylinux_version.py" "${DOCKER_CONTAINER_ID}:/tmp/"
+MANYLINUX_VERSION=$(docker exec "${DOCKER_CONTAINER_ID}" bash -c "/tmp/manylinux_version.py --triton-path ${TRITON_LOCAL_DIR}")
+WHEEL_PLATFORM=manylinux_${MANYLINUX_VERSION}_${WHEEL_ARCH}
+
 docker cp "${SCRIPTS_DIR}/auditwheel_patched.py" "${DOCKER_CONTAINER_ID}:/tmp/"
 docker exec "${DOCKER_CONTAINER_ID}" bash -c "LD_LIBRARY_PATH=${TRITON_LOCAL_DIR}/external_libs /tmp/auditwheel_patched.py -vvvv repair --plat ${WHEEL_PLATFORM} ${WHEEL_PATH}"